Reduce repeated work in forward checking. Diagnostic option added. Work started on guess option.
While forward checking, a direction is associated with each tile,
and a tile in a certain direction will only check a select set
of its neighbors. This is to reduce repeatedly checking many of the
same tiles.

The diagnostic option gets an average value for surrounding tiles
of each blank, and prints if the blank is a mine or not a mine.